add: calorees and details modules, delete: home,onboarding and settings modules
