+++ core/network/src/commonMain/kotlin/com/theophiluskibet/caloree/network/di/NetworkModule.kt
@@ -41,7 +41,7 @@ val networkModule =
                     logger =
                         object : Logger {
                             override fun log(message: String) {
-                                log(message)
+                                // log(message)
                             }
                         }
                     level = LogLevel.ALL
