+++ feature/details/src/commonMain/kotlin/com/theophiluskibet/calorees/details/di/DetailsModule.kt
@@ -1,0 +1,7 @@
+package com.theophiluskibet.calorees.details.di
+
+import org.koin.dsl.module
+
+val detailsModule =
+    module {
+    }
