+++ feature/details/src/commonMain/kotlin/com/theophiluskibet/calorees/details/screens/DetailScreen.kt
@@ -1,0 +1,25 @@
+package com.theophiluskibet.calorees.details.screens
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun DetailScreen(
+    modifier: Modifier = Modifier,
+    name: String,
+) {
+    Column(
+        modifier =
+            Modifier
+                .fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Text("$name")
+    }
+}
