+++ feature/details/src/commonMain/kotlin/com/theophiluskibet/calorees/details/utils/DetailsUiState.kt
@@ -1,0 +1,3 @@
+package com.theophiluskibet.calorees.details.utils
+
+class DetailsUiState
