--- feature/home/src/commonMain/kotlin/com/theophiluskibet/caloree/home/screens/caloreedetails/CaloreeDetails.kt
@@ -1,12 +1,0 @@
-package com.theophiluskibet.caloree.home.screens.caloreedetails
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import org.koin.compose.viewmodel.koinViewModel
-
-@Composable
-fun CaloreeDetails(
-    modifier: Modifier = Modifier,
-    viewModel: CaloreeDetailsViewModel = koinViewModel(),
-) {
-}
