--- feature/home/src/commonMain/kotlin/com/theophiluskibet/caloree/home/screens/caloreedetails/CaloreeDetailsViewModel.kt
@@ -1,8 +1,0 @@
-package com.theophiluskibet.caloree.home.screens.caloreedetails
-
-import androidx.lifecycle.ViewModel
-import com.theophilus.kibet.caloree.data.sources.CaloreeRepository
-
-class CaloreeDetailsViewModel(
-    private val repository: CaloreeRepository,
-) : ViewModel()
