--- feature/onboarding/src/commonMain/kotlin/com/theophiluskibet/caloree/onboarding/Module.kt
@@ -1,5 +1,0 @@
-package com.theophiluskibet.caloree.onboarding
-
-class Module {
-    fun doSomething(): String = "Caloree"
-}
