--- feature/settings/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    alias(libs.plugins.caloreeMultiplaform)
-}
-
-android {
-    namespace = "com.theophiluskibet.caloree.settings"
-}
