--- feature/settings/src/commonMain/kotlin/com/theophiluskibet/caloree/settings/Module.kt
@@ -1,5 +1,0 @@
-package com.theophiluskibet.caloree.settings
-
-class Module {
-    fun doSomething(): String = "Caloree"
-}
