+++ settings.gradle.kts
@@ -34,6 +34,5 @@ include(":core:data")
 include(":core:designsystem")
 include(":core:local")
 include(":core:network")
-include(":feature:onboarding")
-include(":feature:home")
-include(":feature:settings")
+include(":feature:calorees")
+include(":feature:details")
